(SIMP-767) Update to handle RPM macros
The `rpmspec` command that we were using was not present in the native
version of the RPM build tools included in EL6.

As such, this has been replaced with a (much simpler) usage of 'rpm
--specfile'.

While doing this, I ran into issues with RPM macro translation and the
helpers so went ahead and solved that issue as well.

SIMP-767 #comment Required update for the simp-rake-helpers RubyGem
SIMP-224 #close