(SIMP-1181) Build any module as an RPM
This patch adds a Lua-based RPM template to the stack and injects it
into the build process.

By doing this, we can now build *any* Puppet module as an RPM without
forking it.

There must be a well formed metadat.json for this to function properly.

Most importantly, the 'name' field must be '<author>-<module_name>' and
the version number must be intact.

SIMP-1181 #close