(SIMP-1395) Sanity check pkg:rpmsign
Previously the `pkg:rpmsign` task would fail with hundreds of lines of
mostly unhelpful output if the `rpmsign` package was unavailable.  This
adds a sanity check on the executable, raising an exception if it is not
available, and outputing a more helpful message.

Additionally, it updates a similar sanity check in `mock_pre_check()` to
be more robust.

SIMP-1395 #close

Change-Id: I26cabace706d2f2d23a01143bab95e5b197f18c1