Update patch set 3
Patch Set 3: Code-Review-2

So, I'm OK with this, but the Rakefiles should already have the capability to check the files in the module for changes and prevent building if the tar file is newer than any given changes to the associated files. We may simply need to exclude some files from the list that is being checked.

The method in question if FileUtilis.uptodate?

Patch-set: 3
Reviewer: Trevor Vaughan <1004834@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=-2