(SIMP-1423) Add a logger and logging command runner
This patch adds a logger via the Logging gem.  Previously there was no
such thing, with log-like messages going directly to the console
(stdout). While this patch does NOT capture such streams, it sets up the
some infrastructure to do so.  Namely, it sets up a `Logging::Logger`
for the `Simp` module and sub-loggers for most sub-modules and
sub-classes.  All logs are aggregated into two streams: a file that
collects all logs, and a stderr stream that filters out all messages
except those of severity "fatal," "error," or "warn."

Additionaly, this patch includes a command runner that captures stdout
and stderr streams of the sub-process and redirects them into the
logger.  For now it classifies all stdout as "debug" messages and all
stderr output as "error" messages, but this could be made configurable
in a future patch.  The runner uses `Open3.popen3` and returns a
`Process::Waiter` so it should be thread-safe.

SIMP-1423 #close