(SIMP-3702) simp-rake-helpers Tweak list of required commands
Removed rpmkeys and rpmspec from the required list of commands,
as they are not present in CentOS 6 and are automatically
installed when the required rpmbuild command is installed in
CentOS 7.

Long term solution should probably be a list of RPM packages,
per OS, instead of a single list of individual commands.

SIMP-3702 #close