(SIMP-4957) Fix regression in `SIMP_BUILD_distro`
A recent commit inadvertently removed the ability for the
`SIMP_BUILD_distro` to specify a specific build target for SIMP by
hard-coding each value to the Facter facts of the build system.

This patch reintroduces support for `SIMP_BUILD_distro` and falls back
to the Facter values if it isn't set.

SIMP-4957 #close