(SIMP-5850) simp-adapter fails to cleanup
This fix, along with a corresponding one in simp-adapter,
addresses a problem in which, during upgrade of a SIMP
component RPM, files only in the old version were not deleted
appropriately by simp_rpm_helper in the directory copy.

The fix is to move the simp_rpm_helper rsync operation from
the %post into the %posttrans, along with a marker file to
indicate when it should be run.  When simp_rpm_helper executed
the rsync in the %post, it was copying the files *before* RPM
had removed the old package files.

SIMP-5850 #comment simp-rake-helper fixes