(SIMP-9748) Fix `not a valid ISO` error on EL8
Before this patch, running `rake unpack` on a valid ISO file from the
EL8 build container would fail with the error, `Error: The file provided
is not a valid ISO.`  It seems that on EL8, the expected `ISO 9660`
string is no longer returned by the file command as the first match.

This patch fixes the issue by running `file --keep-going` to return ALL
matches, which still include the expected `ISO 9660` string.

[SIMP-9748] #close